+++ assets/js/aione.js
@@ -31275,9 +31275,35 @@ function init_sliders(){
         $.each(aione_slider_ids, function( index, aione_slider ) {
             var slider_data = $('#'+aione_slider).data();
             if(slider_data !== undefined){
+
+                console.log('Slider Data =');
+                console.log(slider_data);
+                /*
+                $.each( slider_data, function( key, value ) {
+                    is_json = true;
+                    try {
+                        JSON.parse( value );
+                    } catch (e) {
+                        is_json = false;
+                    }
+
+                    if( is_json ){
+                        console.log("is_json");
+                        slider_data[key] = JSON.parse( value ); 
+                    }
+                });
+                */
+
+                var desktop_items = slider_data.items;
+                var tablet_items = parseInt( desktop_items / 2 );
+                var mobile_items = parseInt( tablet_items / 2 );
+
+                slider_data.responsive = {0:{items:mobile_items},601:{items:tablet_items},901:{items:desktop_items}};
+                console.log(slider_data);
+
                 $('#'+aione_slider).owlCarousel(slider_data);
             } else {
-                console.log('Dlider Data =' + slider_data);
+                console.log('Slider Data =' + slider_data);
                 $('#'+aione_slider).owlCarousel({
                     items:1,
                     loop:true,