+++ CHANGELOG.md
@@ -187,3 +187,6 @@ ADDED: Scroll Progress Bar
 ## 7.4.0.5 04-12-2019
 ADDED: Google Analytics Code
 FIXED: Per post setting js conflict
+
+## 7.4.0.6 10-12-2019
+ADDED: Responsive Slider